代码重构

+++ src/main/java/org/snaker/engine/scheduling/JobEntity.java
@@ -62,19 +62,19 @@ public class JobEntity implements Serializable {
 	 */
 	private Date startTime;
 	/**
-	 * 间隔时间
+	 * 间隔时间(分钟)
 	 */
-	private long period;
+	private int period;
 	/**
 	 * 执行参数
 	 */
 	private Map<String, Object> args;
 	
 	public JobEntity(String id, Task task, Date startTime) {
-		this(id, task, startTime, 0L);
+		this(id, task, startTime, 0);
 	}
 	
-	public JobEntity(String id, Task task, Date startTime, long period) {
+	public JobEntity(String id, Task task, Date startTime, int period) {
 		this.id = id;
 		this.task = task;
 		this.startTime = startTime;
@@ -82,10 +82,10 @@ public class JobEntity implements Serializable {
 	}
 	
 	public JobEntity(String id, Task task, Date startTime, Map<String, Object> args) {
-		this(id, task, startTime, args, 0L);
+		this(id, task, startTime, args, 0);
 	}
 	
-	public JobEntity(String id, Task task, Date startTime, Map<String, Object> args, long period) {
+	public JobEntity(String id, Task task, Date startTime, Map<String, Object> args, int period) {
 		this.id = id;
 		this.task = task;
 		this.startTime = startTime;
@@ -105,10 +105,10 @@ public class JobEntity implements Serializable {
 	public void setStartTime(Date startTime) {
 		this.startTime = startTime;
 	}
-	public long getPeriod() {
+	public int getPeriod() {
 		return period;
 	}
-	public void setPeriod(long period) {
+	public void setPeriod(int period) {
 		this.period = period;
 	}
 	public Map<String, Object> getArgs() {
